Refactor: Move API parameter and response retrieval functions to a new utility module
